Create Widgets and Gizmos
An online retailer sells two products: widgets and gizmos. Each widget weighs 75 grams. Each gizmo weighs 112 grams. Write a program that reads the number of
widgets and the number of gizmos in an order from the user. Then your program should compute and display the total weight of the order.
TEST CASE 1

INPUT
28
22
OUTPUT
The number of widgest is 28.0 and the number of gizmos is 22.0 and the total weight is 4.564 Kilograms
TEST CASE 2

INPUT
86
47
OUTPUT
The number of widgest is 86.0 and the number of gizmos is 47.0 and the total weight is 11.714 Kilograms