Create Conversion
he decimal or denary counting system uses the Base-of-10 numbering system where each digit in a number takes on one of ten possible values, called digits, from 0 to 9, eg. 21310 (Two Hundred and Thirteen).

But as well as having 10 digits ( 0 through 9 ), the decimal numbering system also has the operations of addition ( + ), subtraction ( ), multiplication ( ) and division ( ).

In a decimal system each digit has a value ten times greater than its previous number and this decimal numbering system uses a set of symbols, b, together with a base, q, to determine the weight of each digit within a number. For example, the six in sixty has a lower weighting than the six in six hundred. Then in a binary numbering system we need some way of converting Decimal to Binary as well as back from Binary to Decimal.
TEST CASE 1

INPUT
4
OUTPUT
100
TEST CASE 2

INPUT
521
OUTPUT
1000001001