Create BUILDING
Create a class called Building with instance "objname" with the following private data members or member variables

length, width and ratePerSqFeet (all of type int).

Include the following public member functions or methods:

initializeData() that accepts 3 arguments and its return type is void. This function is used to set the values for length, width and ratePerSqFeet. The arguments are passed to this function in the
same order.

getLength() --- to return the length of the building

getWidth() --- to return the width of the building

getRatePerSqFeet() --- to return the ratePerSqft of the building

calculateCost() --- to calculate and return the cost of the building

determineSuitability() --- to determine and return the type of the building.

Mandatory:

You should call the following methods from the main class (TestClass)

getRatePerSqFeet and determineSuitability as follows

objname.getRatePerSqFeet()

objname.determineSuitability()

We are interested in Buildings that are almost in the shape of a square. If the length and width of the building differ by atmost 10, then the building is suitable. If the difference between the length and width of the building is more than 10, then it is not suitable. In the main function / method, create an object of type buiding and invoke the corresponding methods.
TEST CASE 1

INPUT
34
56
54
OUTPUT
Length : 34
Width : 56
Rate Per SqFt : 54
Cost of the Building : 102816
Suitability : Not Suitable
TEST CASE 2

INPUT
400
395
1000
OUTPUT
Length : 400
Width : 395
Rate Per SqFt : 1000
Cost of the Building : 158000000
Suitability : Suitable